Fixing Travis CI build matrix expansion
It seems like the CI build matrix does not expand correctly
when an `os` is not specified, so I've set that to `linux`.

Out of caution, the current build job's ubuntu distribution
is also passed into `rosdep`. It seems to resolve the missing
dependencies when said dependencies clearly exist for a given
distribution.

Finally, removing `sudo: true` because all jobs have su
privilege by default now.

Signed-off-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>